Refactor CharacterPartAttachment: make a system
