+++ Assets/Scripts/Field/Character/Character.cs
@@ -146,7 +146,7 @@ public class Character : MonoBehaviour
                 {
                     _mainPart = characterPart;
                     ability.Apply();
-                    var newMainPart = _mainPart.CharacterPartAttachment.DetachParts(_mainPart);
+                    var newMainPart = _mainPart.CharacterPartAttachment.DetachPartsAndUnite(_mainPart);
 
                     //if (result)
                     //    Detached.Invoke();
