+++ Assets/Scripts/Field/Character/CharacterPart.cs
@@ -49,8 +49,6 @@ public class CharacterPart
 
     
     private Field _field;
-    public CharacterPartAttachment CharacterPartAttachment;
-
 
     public CharacterPart()
     {
