--- Assets/Scripts/Field/Character/CharacterPartAttachment.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: aa670df5ad25490d8a79f682f4b35f6c
-timeCreated: 1655113624
