+++ Assets/Scripts/Systems/AttachmentSystem.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7b742b73f46f4474b59337ea6772e19c
+timeCreated: 1658406732
